Changed lastUpdateEpoch field type's from Int to Long.

+++ app/src/main/java/android/azadev/weatherry/data/remote/response/WeatherCurrent.kt
@@ -23,7 +23,7 @@ data class WeatherCurrent(
     @SerializedName("last_updated")
     val lastUpdated: String,
     @SerializedName("last_updated_epoch")
-    val lastUpdatedEpoch: Int,
+    val lastUpdatedEpoch: Long,
     @SerializedName("precip_in")
     val precipIn: Double,
     @SerializedName("precip_mm")
